pyproject.toml
#
# pyproject.toml
#

[project]
name = "provide-foundation"
description = "Foundation Telemetry: An opinionated, developer-friendly telemetry wrapper for Python."
dynamic = ["version"]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Tim Perkins", email = "code@tim.life" },
]
maintainers = [
    { name = "provide.io", email = "code@provide.io" },
]
keywords = ["telemetry", "logging", "tracing", "python", "pyvider"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Logging",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "aiofiles>=23.2.1",
    "attrs>=23.1.0",
    "structlog>=25.3.0",
    "tomli_w>=1.0.0",
]

# [project.urls]
# Homepage = "https://foundation.provide.io"
# Documentation = "https://foundation.provide.io"
# Repository = "https://github.com/provide-io/provide-foundation"
# Issues = "https://github.com/provide-io/provide-foundation/issues"
# Examples = "https://github.com/provide-io/provide-foundation/tree/main/examples"


[project.optional-dependencies]
cli = [
    "click>=8.1.7",
]
crypto = [
    "cryptography>=45.0.7",
]
opentelemetry = [
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22.0",
    "opentelemetry-exporter-otlp-proto-http>=1.22.0",
]
all = [
    "provide-foundation[cli,crypto,opentelemetry]",
]

[project.scripts]
foundation = "provide.foundation.cli.main:cli [cli]"

[dependency-groups]
dev = [
    "bandit>=1.8.3",
    "hypothesis>=6.131.28",
    "mypy>=1.8.0",
    "psutil>=7.0.0",
    "pyrefly>=0.16.2", "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.15.0",
    "pytest-xdist>=3.3.0",
    "pyyaml>=6.0.2",
    "reuse>=1.1.0",
    "ruff>=0.1.0",
    "twine>=6.1.0",
    "ty>=0.0.1a6",
    "types-cryptography>=3.3.23.2",
    "types-toml>=0.10.8.20240310",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
    "mike>=2.0.0",
]

################################################################################
# Pytest Configuration
################################################################################

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"

asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

testpaths = ["tests"]
pythonpath = ["src", "."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "*Tests"]
python_functions = ["test_*", "*_test"]
markers = [
    "serial: run tests serially to avoid conflicts on global state",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:cannot collect test class 'Test.*' because it has a __init__ constructor:pytest.PytestCollectionWarning",
    "ignore:cannot collect test class .* because it has a __init__ constructor:pytest.PytestCollectionWarning",
    "ignore:.* KqueueSelector constructor:pytest.PytestCollectionWarning",
    "ignore:.* KqueueSelector constructor:pytest.PytestCollectionWarning",
]
norecursedirs = [
    ".git", ".hg", ".svn", "*_build", "build", "dist", "*.egg-info",
    ".venv", "venv", # Added venv as it's a common default
    "htmlcov", "docs/_build",
    ".hypothesis", # Added to address the warning
]
addopts = "-ra -q --color=yes --benchmark-columns=min,max,mean,stddev,median,iqr,ops --benchmark-sort=mean"

################################################################################
# Ruff Linter and Formatter Configuration
################################################################################

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "F", "W", "I", "UP", "ANN", "B", "C90", "SIM", "PTH", "RUF",
]
ignore = [ "ANN401", "B008", "E501" ] # Removed ANN101, ANN102
[tool.ruff.lint.isort]
known-first-party = ["pyvider", "tests"]
force-sort-within-sections = true
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

################################################################################
# MyPy Static Type Checker Configuration
################################################################################

[tool.mypy]
mypy_path = "src"
# Show error codes in output.
show_error_codes = true
# Show column numbers in output.
show_column_numbers = true
# Use visually nicer output in error messages:
pretty = true
# Use current Python version.
python_version = "3.11"
# Warn about unused '# type: ignore' comments.
warn_unused_ignores = true
# Warn about unused '[mypy-pattern]'  config sections.
warn_unused_configs = true
# Do not follow imports to .pyi stubs for external libraries.
follow_imports = "normal"  # Changed global strategy
# Do not check that test functions are annotated.
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = [
    "structlog",
    "structlog.*"
]
follow_imports = "skip" # Keep structlog skipped
ignore_missing_imports = true

################################################################################
# Coverage.py Configuration
################################################################################

[tool.coverage.run]
source = ["provide.foundation"]
branch = true
parallel = true

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80 # Set this to a realistic target once you get >0%
precision = 2
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "def __repr__",
    "def __str__",
    "@(abc\\.)?abstractmethod",
    "pass",
    # This is usually fine, but double-check it's not too greedy for your codebase
    "logger\\.(debug|info|warning|error|exception|critical|trace)\\(",
    "import logging", # Excluding all import logging lines might be too much
    "from logging import", # Same as above
    "print\\(",
]

[tool.coverage.html]
directory = "htmlcov" # Using an absolute path

[tool.coverage.xml]
output = "coverage.xml"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.bandit]
exclude_dirs = [".venv", "tests"]

[tool.pyrefly]
project_includes = [
    "src",
    "tests",
]

# 🐍🏗️⚙️
